requirements.txt
pvporcupine
pvrecorder
winsound
pygame
openai
python-dotenv
